Starting to move functionality to move granules
