+++ tasks/move-granules/package.json
@@ -1,0 +1,56 @@
+{
+  "name": "@cumulus/move-granules",
+  "version": "1.5.0",
+  "description": "Move granule files from staging to final location",
+  "main": "index.js",
+  "directories": {
+    "test": "tests"
+  },
+  "homepage": "https://github.com/cumulus-nasa/cumulus/tree/master/tasks/move-granules",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/cumulus-nasa/cumulus"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "engine": {
+    "node": ">=8.10.0"
+  },
+  "scripts": {
+    "test": "ava",
+    "test-coverage": "nyc ava",
+    "build": "rm -rf dist && mkdir dist && cp -R schemas dist/ && webpack",
+    "watch": "rm -rf dist && mkdir dist && cp -R schemas dist/ && webpack --progress -w",
+    "prepublishOnly": "PRODUCTION=true npm run build"
+  },
+  "ava": {
+    "files": "tests",
+    "serial": true,
+    "color": false
+  },
+  "nyc": {
+    "exclude": [
+      "tests"
+    ]
+  },
+  "author": "Cumulus Authors",
+  "license": "Apache-2.0",
+  "dependencies": {
+    "@cumulus/cmrjs": "^1.5.0",
+    "@cumulus/common": "^1.5.0",
+    "@cumulus/cumulus-message-adapter-js": "^1.0.1",
+    "@cumulus/ingest": "^1.5.0",
+    "@cumulus/test-data": "^1.4.0",
+    "lodash.get": "^4.4.2",
+    "xml2js": "^0.4.19"
+  },
+  "devDependencies": {
+    "ava": "^0.25.0",
+    "nyc": "^11.6.0",
+    "proxyquire": "^2.0.0",
+    "sinon": "^4.5.0",
+    "webpack": "~4.5.0",
+    "webpack-cli": "~2.0.14"
+  }
+}
