+++ tasks/move-granules/schemas/output.json
@@ -1,0 +1,48 @@
+{
+  "title": "MoveGranulesOutput",
+  "description": "Describes the output produced by the move-granules task",
+  "type": "object",
+  "properties": {
+    "allGranules": {
+      "type": "object",
+      "description": "Object of all granules where granuleID is the key",
+      "properties": {
+        "granuleId": { "type": "string" }
+      }
+    },
+    "cmrFiles": {
+      "type": "array",
+      "description": "list of CMR files from input"
+    },
+    "granules": {
+      "type": "array",
+      "items": {
+        "type": "object",
+        "required": ["granuleId", "files"],
+        "properties": {
+          "granuleId": {
+            "type": "string"
+          },
+          "files": {
+            "type": "array",
+            "items": {
+              "type": "object",
+              "required": ["name"],
+              "properties": {
+                "name": {
+                  "type": "string"
+                },
+                "filename": {
+                  "type": "string"
+                },
+                "bucket": {
+                  "type": "string"
+                }
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+}
