+++ tasks/move-granules/webpack.config.js
@@ -1,0 +1,25 @@
+const path = require('path');
+
+let mode = 'development';
+let devtool = 'inline-source-map';
+
+if(process.env.PRODUCTION) {
+  mode = 'production',
+  devtool = false  
+}
+
+module.exports = {
+  mode,
+  entry: './index.js',
+  output: {
+    libraryTarget: 'commonjs2',
+    filename: 'index.js',
+    path: path.resolve(__dirname, 'dist')
+  },
+  externals: [
+    'aws-sdk',
+    'electron'
+  ],
+  devtool,
+  target: 'node'
+};
