+++ tasks/post-to-cmr/schemas/config.json
@@ -5,40 +5,16 @@
   "required": [
     "bucket",
     "stack",
-    "cmr",
-    "input_granules",
-    "collection",
-    "buckets"
+    "cmr"
   ],
   "properties": {
-    "input_granules": {
-      "type": "array",
-      "description": "granules object used to construct output for cumulus indexer",
-      "items": {
-        "type": "object"
-      }
-    },
     "bucket": {
       "type": "string",
       "description": "the bucket the has the private/public key needed for decrypting cmr password"
     },
-    "buckets": {
-      "type": "object",
-      "description": "aws s3 buckets used by this task",
-      "properties": {
-        "internal": { "type": "string" },
-        "private": { "type": "string" },
-        "protected": { "type": "string" },
-        "public": { "type": "string" }
-      }
-    },
     "stack": {
       "type": "string",
       "description": "The name of the deployment stack"
-    },
-    "granuleIdExtraction": {
-      "type": "string",
-      "description": "The regex needed for extracting granuleId from filenames"
     },
     "cmr": {
       "type": "object",
@@ -49,31 +25,6 @@
         "clientId": { "type": "string" },
         "username": { "type": "string" },
         "password": { "type": "string" }
-      }
-    },
-    "collection": {
-      "type": "object",
-      "required": [
-        "files",
-        "name"
-      ],
-      "properties": {
-        "name": { "type": "string" },
-        "process": { "type": "string" },
-        "url_path": { "type": "string" },
-        "duplicateHandling": { "type": "string" },
-        "files": {
-          "type": "array",
-          "items": {
-            "type": "object",
-            "required": ["regex", "bucket"],
-            "properties": {
-              "regex": { "type": "string" },
-              "bucket": { "type": "string" },
-              "url_path": { "type": "string" }
-            }
-          }
-        }
       }
     }
   }
